chore: :hammer: lvl2 refactor
