+++ src/blueprints/helper.py
@@ -1,0 +1,10 @@
+def get_currency_symbol(currency):
+    currency_to_symbol_dict = {
+        "INR" : "₹",
+        "USD" : "$ ",
+        "EUR" : "€ ",
+        "GBP" : "£ ",
+        "JPY" : "¥ ",
+        "AUD" : "A$ ",
+    }
+    return currency_to_symbol_dict.get(currency, currency)
