+++ src/static/css/other_calculators.css
