+++ src/static/css/settings.css
