+++ src/templates/admin_dashboard.html
@@ -44,11 +44,11 @@
     <div class="admin-tools">
         <h3><i class="fas fa-tools"></i> Admin Tools</h3>
         <div class="button-grid">
-            <a href="{{ url_for('manage_users') }}" class="btn btn-large"><i class="fas fa-user-cog"></i> User Management</a>
-            <a href="{{ url_for('history') }}" class="btn btn-large"><i class="fas fa-history"></i>Transaction History</a>
-            <a href="{{ url_for('settings') }}" class="btn btn-large"><i class="fas fa-cog"></i> System Settings</a>
-            <a href="{{ url_for('audit_log') }}" class="btn btn-large"><i class="fas fa-clipboard-list"></i> Audit Logs</a>
-            <a href="{{ url_for('update_jeweller_details') }}" class="btn btn-large"><i class="fas fa-cog"></i>Jewellers Config Settings</a>
+            <a href="{{ url_for('admin.manage_users') }}" class="btn btn-large"><i class="fas fa-user-cog"></i> User Management</a>
+            <a href="{{ url_for('history.history') }}" class="btn btn-large"><i class="fas fa-history"></i>Transaction History</a>
+            <a href="{{ url_for('admin.settings') }}" class="btn btn-large"><i class="fas fa-cog"></i> System Settings</a>
+            <a href="{{ url_for('admin.audit_log') }}" class="btn btn-large"><i class="fas fa-clipboard-list"></i> Audit Logs</a>
+            <a href="{{ url_for('config.update_jeweller_details') }}" class="btn btn-large"><i class="fas fa-cog"></i>Jewellers Config Settings</a>
         </div>
     </div>
     <div class="admin-dashboard">
