+++ src/templates/config.html
@@ -6,7 +6,7 @@
 {% block content %}
 <div class="container">
     <h1>Jeweller Details</h1>
-    <form method="POST" action="{{ url_for('update_jeweller_details') }}" enctype="multipart/form-data">
+    <form method="POST" action="{{ url_for('config.update_jeweller_details') }}" enctype="multipart/form-data">
         <div class="row">
             <div class="col-md-6">
                 <div class="form-group">
