+++ src/templates/ext/footer.html
