+++ src/templates/ext/navbar.html
@@ -1,5 +1,5 @@
 <nav class="navbar navbar-expand-lg navbar-custom">
-    <a class="navbar-brand" href="{{ url_for('home') }}">
+    <a class="navbar-brand" href="{{ url_for('homepage.home') }}">
         <i class="fas fa-home"></i>
         Home
     </a>
@@ -11,23 +11,23 @@
         <ul class="navbar-nav ml-auto">
             {% if current_user.is_authenticated %}
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('dashboard') }}">
+                    <a class="nav-link" href="{{ url_for('admin.dashboard') }}">
                         <i class="fas fa-tachometer-alt"></i> Dashboard
                     </a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('logout') }}">
+                    <a class="nav-link" href="{{ url_for('auth.logout') }}">
                         <i class="fas fa-sign-out-alt"></i> Logout
                     </a>
                 </li>
             {% else %}
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('login') }}">
+                    <a class="nav-link" href="{{ url_for('auth.login') }}">
                         <i class="fas fa-sign-in-alt"></i> Login
                     </a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('register') }}">
+                    <a class="nav-link" href="{{ url_for('auth.register') }}">
                         <i class="fas fa-user-plus"></i> Sign Up
                     </a>
                 </li>
