+++ src/templates/homepage.html
@@ -10,19 +10,19 @@
     <p>Select a calculator to get started.</p>
     <div class="hero-buttons">
         {% if is_gold_calculator_enabled %}
-            <a href="{{ url_for('gold_calculator') }}" class="btn btn-hero">
+            <a href="{{ url_for('gold_calculator.gold_calculator') }}" class="btn btn-hero">
                 <i class="fa-solid fa-coins fa-lg"></i> Gold Calculator
             </a>
         {% endif %}
         {% if is_silver_calculator_enabled %}
-        <a href="{{ url_for('silver_calculator') }}" class="btn btn-hero">
+        <a href="{{ url_for('silver_calculator.silver_calculator') }}" class="btn btn-hero">
             <i class="fa-solid fa-gem fa-lg"></i> Silver Calculator
         </a>
         {% endif %}
-        <a href="{{ url_for('features') }}" class="btn btn-hero">
+        <a href="{{ url_for('additional.features') }}" class="btn btn-hero">
             <i class="fa-solid fa-medal fa-lg"></i> Features
         </a>
-        <a href="{{ url_for('pricing') }}" class="btn btn-hero">
+        <a href="{{ url_for('additional.pricing') }}" class="btn btn-hero">
             <i class="fa-solid fa-money-check-dollar fa-lg"></i> Pricing
         </a>
     </div>
