+++ src/templates/settings.html
@@ -6,7 +6,7 @@
 {% block content %}
 <div class="container">
     <h1>System Settings</h1>
-    <form method="POST" action="{{ url_for('settings') }}">
+    <form method="POST" action="{{ url_for('admin.settings') }}">
         <div class="row">
             <div class="col-md-6">
                 <div class="form-group">
